flume sink: Fix jar shading
Previously, when generating the kudu-flume-sink jar, we pulled in many
unshaded dependencies. This change improves the situation considerably
by doing the following:

1. Make the hadoop jars provided. This is fine because Flume always puts
   the Hadoop jars for the version it is running against on its
   classpath.
2. Use the transitive shading provided by the kudu-client jar. This
   makes it so that the kudu-client dependencies don't pollute the
   classpath. Previously, unshaded kudu-client dependencies were being
   included in the kudu-flume-sink jar (which I found surprising).

After this change the only non-org.apache.kudu classes in the classpath
are inherited from kudu-client, and once we fix KUDU-1780 then that will
be resolved as well.

Change-Id: I0ac2b67636c9f13f56cec2ea526d42c9e11f6b78
Reviewed-on: http://gerrit.cloudera.org:8080/5327
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>
Tested-by: Kudu Jenkins