KUDU-766: limit number of glog files
This commit introduces a new flag, 'max_log_files', that limits the
maximum number of rolled glog files to keep around at each severity
level. This flag matches an Apache Impala flag of the same name, and the
implementation is partly borrowed from it.

I initially went the route of making a new maintenance manager operation
type to schedule the file cleanup, but it proved difficult to shoehorn
this usecase into that abstraction.  Instead, a thead is spawned on
startup that checks for excess files every 60 seconds.

Change-Id: Ia3f81638a6dfcf134665322f091ff26d29640fd9
Reviewed-on: http://gerrit.cloudera.org:8080/5340
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>