KUDU-1753 [delete_table-test] deleted-while-in-scan test
Added an integration test to ensure a tablet server keeps all the
necessary data around until a scan operation ends even if tablet
is being deleted concurrently.

The test is disabled as it currently fails. A follow up patch will fix
the bug and enable the test.

This is in the context of the following JIRA item:
KUDU-1753 Impala query fails: Unable to advance iterator:
  Illegal state: Tablet is not running

Change-Id: I983e68862b5535f2f95eedd41850a8a88e95e69c
Reviewed-on: http://gerrit.cloudera.org:8080/5345
Tested-by: Kudu Jenkins
Reviewed-by: David Ribeiro Alves <dralves@apache.org>