rpc-test: fix flakiness of TestCallLongerThanKeepalive
The 50ms keepalive was apparently short enough that occasionally the
server connection would drop before even receiving the initial call on
the connection. This would cause the test to fail with a spurious
"connection reset" type error.

I looped the following to test:

/build-support/dist_test.py --disable-sharding \
   loop -n 1000 build/latest/bin/rpc-test \
   --gtest_filter=OptionalSSL/TestRpc.TestCallLongerThanKeepalive/1 \
   --stress_cpu_threads=8

Without the patch[1] there were 31/1000 failures, all due to this issue.
With the patch[2] there were no failures.

[1] http://dist-test.cloudera.org//job?job_id=todd.1480911578.21798
[2] http://dist-test.cloudera.org//job?job_id=todd.1480911736.2857

Change-Id: I8d8d981af6dae90d0c656745d9260ffa920c9039
Reviewed-on: http://gerrit.cloudera.org:8080/5355
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins