client: don't log retriable scanner errors
Retriable scanner errors are a normal part of life. Prior to this patch,
every such error (eg SERVER_TOO_BUSY) resulted in both an INFO and a
WARNING message logged in the client.

In an Impala stress scenario this was producing several hundred logs per
second, even while queries were mostly succeeding.

This downgrades them to a VLOG(1) level.

Change-Id: I32e16dd10f9de8a015d420da4b58b2f363045f67
Reviewed-on: http://gerrit.cloudera.org:8080/5380
Tested-by: Kudu Jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>
Reviewed-by: Adar Dembo <adar@cloudera.com>