Make delete_table-test less flaky
This patch reduces the flakiness of delete_table-test by solving
flakiness problems on two individual tests:

1. DeleteTableTest.TestAutoTombstoneAfterTabletCopyRemoteFails

This test triggers a remote crash that would sometimes prevent an RPC
response from arriving at the caller. We now account for this.

2. DeleteTableTombstonedParamTest.TestTabletTombstone

This test would sometimes not write quickly enough, resulting in a
timeout waiting for 3 wal files to appear on tablet 0. This could occur
when running under TSAN on a system under heavy load.

Now, the above two tests run successfully in 200/200 attempts on TSAN
under load.

This patch does not address the flakiness of
DeleteTableTest.TestDeleteTableWithConcurrentWrites yet. That test has
two issues that are being tracked as KUDU-1294 and KUDU-1797.

However a log message has been added to this commit to aid in debugging
that test in a later patch.

Change-Id: Iea7760009acc8a795225369721af97583974feba
Reviewed-on: http://gerrit.cloudera.org:8080/5421
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>