Make fuzz-itest run serially
This test relies on restarting servers, and on the flaky dashboard it
appears to fail often due to port conflicts. Rather than fix it at the
moment, let's just make it run serially. This won't affect precommit
time, since in the dist-test environment, _all_ tests run alone on a
machine.

Change-Id: Ia6e44b4318fec14f6f0cbc8e697fe39049a6b246
Reviewed-on: http://gerrit.cloudera.org:8080/5398
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>