csd: two minor changes
Revert the memory limit scaleFactor to the default value (1.0). 1.3 is
recommended for Java processes; not sure why Kudu was using that.

Convert the core dump directory to two separate role-level parameters. I
relented and did this once I saw that Impala also defines its core dump
directory in this way. In testing, the conversion was fine when the
service-level parameter wasn't set. When it was set, I saw that the
overridden value carried over to the roles even after the conversion
(good!). However, attempting to remove the so-called "role" overrides in the
UI yielded an error (bad!).

To work around this, I changed the names of the parameters (i.e. I prepended
either 'master' or 'tserver' to the names). This means any old value is lost
during an upgrade to the new CSD (a little bad), but that's better than not
being able to remove the value altogether.

Change-Id: I74b2fcb69fde9e0dbb232dba78db53cc1543b489
Reviewed-on: http://gerrit.cloudera.org:8080/5435
Tested-by: Kudu Jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>
(cherry picked from commit 251860b36fae96807c625f54d96c67e578685ab4)
Reviewed-on: http://gerrit.cloudera.org:8080/5450