MiniKuduClusterBuilder.build should clean up state when initialization fails
The MiniKuduCluster starts processes and does many other potentially
failing operations as part of the constructor. This change moves these
potentially failing operations out of the constructor so that the
MiniKuduCluster can be shutdown cleanly on partial-initialization
failure. This change should be completely transparent to users of the
public API.

https://lists.apache.org/thread.html/922614e80f3f64be885f67380bf44730132896dc21480b28d0a93108@%3Cdev.kudu.apache.org%3E
contains more discussion.

Change-Id: I024872cddf352a9b75dab87af685ca3af50099c7
Reviewed-on: http://gerrit.cloudera.org:8080/5448
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins