docs: KUDU-1767. Document possible client op reordering
This patch documents the possibility of client reordering when using
concurrent flush modes in both the C++ and Java clients.

Also adds a note to the "Transaction Semantics" section of the user
guide.

Tested the docs rendering by building the user guide and the C++ and
Java API docs locally.

Also updated some old issue URLs to point to the ASF JIRA instance.

Change-Id: I65215d833c65e54fcf080d61adc5f6ed3d303224
Reviewed-on: http://gerrit.cloudera.org:8080/5464
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Todd Lipcon <todd@apache.org>