external_mini_cluster: check for leaks before killing servers
This improves the ExternalMiniCluster when running with LeakSanitizer
enabled so that, before killing subprocesses, it runs a leak check.
Without this, it's hard for us to know if there are any leaks in code
paths that are covered only by EMC-based tests.

Tested by manually inserting a leak into the tablet server and verifying
that an EMC test failed.

Change-Id: I9b6c38281e35e8300e849f92e91d6955a8d735bf
Reviewed-on: http://gerrit.cloudera.org:8080/5501
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>