TLS-negotiation [1/n]: deprecate unused SaslAuth fields
This commit deprecates two fields in the SaslMessagePB.SaslAuth type:

method: this field has been unused since 0.6. In 0.6 it was a required
field, so this change technically breaks compat with 0.6 clients or
servers, but I don't think we have or guarantee compat with pre 1.0.

challenge: this field existed to allow the server to piggy back a
challenge token along with the NEGOTIATE response, but we never used it.
The Java client currently doesn't handle it, so we could not start using
it in the future in a backwards compatible way, thus it's best to
simplify things by removing it.

Change-Id: I5e40e6453892a78b293f3dd63980eab839b671c1
Reviewed-on: http://gerrit.cloudera.org:8080/5755
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>