KUDU-1812. Redaction support for protobufs
This adds a new local protobuf option 'kudu.REDACT' which can be used to
mark a field as containing user data and thus needing redaction during
stringification.

New 'DebugString' and 'ShortDebugString' variants are provided which
respect the redaction option and use the KUDU_REDACT macro to implement
it on fields tagged with this option.

This patch includes a test protobuf using this feature and also tags the
relevant fields in our various protobufs. However, it does not yet
change call sites to use the SecureDebugString variants -- that is done
in a follow-on commit.

This commit also does not implement redaction for the JSONWriter-based
stringification. As far as I am aware, we do not use JSON output in any
log messages, and we are not yet aiming to redact /tracing.html output,
etc.

Change-Id: Ie07bfdcbc33d38d55315faae2c4906899a5450fb
Reviewed-on: http://gerrit.cloudera.org:8080/5553
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>