KUDU-699 followup: fix flakiness of raft_consensus_quorum-test
The fix for KUDU-699 introduced flakiness in this test. Now that
Peer::Close() doesn't block on outstanding requests, it's possible for
the test proxy's threadpool to get shut down while a request is
outstanding. This means that we now expect to sometimes get
"ServiceUnavailable" when submitting the response callbacks.

Before this patch, looping the test in DEBUG mode with 4 stress CPU
threads failed almost every task. Now it passed 100/100.

Change-Id: Id1c72d98b59b3c5a33b0a2d236a7c77b1ef1b9e3
Reviewed-on: http://gerrit.cloudera.org:8080/5542
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>