Fix flakiness in ToolTest.TestRemoteReplicaCopy
This test created a table, and then immediately asserted on the error
message returned when trying to copy a tablet on top of an existing
tablet. In fact, the destination tablet might still be in the 'creating'
transition, in which case the copy attempt failed with an unexpected
error message.

This patch fixes the issue by waiting for the RUNNING state first.

Tested a loop 500x with 4 stress threads in a RELEASE build (where this
seemed to be most flaky according to the dashboard).

Change-Id: I740f6b469561a17792a315caf4a10c2627b55149
Reviewed-on: http://gerrit.cloudera.org:8080/5543
Tested-by: Kudu Jenkins
Reviewed-by: Dinesh Bhat <dinesh@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>