KUDU-1812. Replace PB DebugString calls with redactable variants
This uses the clang tool from the prior commit to replace all call sites
of Message::DebugString and Message::ShortDebugString with corresponding
SecureDebugString(msg) and SecureShortDebugString(msg) calls.

The commit was done using the tool except for the following:
- a few call sites inside macros weren't handled by the tool, which
  instead inserted 'TODO(PBString)' comments. I fixed those few cases by
  hand.
- re-wrapped to avoid long lines called out by ilint
- added appropriate #includes for the new calls

The only potentially controversial bit here is whether we should make
this change in the various tests. In fact, the tests that are checking
whether one PB matches another should probably not be redacting
anything. However, the tests also configure redaction to be disabled, so
the secure and regular variants should have identical output. I chose to
make the substitutions in test code as well as production code following
the reasoning that it is less cognitive load (and easier to check in
precommit) to say "thou shalt never use Message::DebugString" rather
than have different rules in the case of tests.

Change-Id: I2c5d1355bdfdbf2232aae8c0d809cc044790de28
Reviewed-on: http://gerrit.cloudera.org:8080/5562
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>