EMC: Don't reuse data dir for log dir
Separate the data and log directories in ExternalMiniCluster. Prior to
this patch, the ExternalMiniCluster often used a shared directory for
the data dir and the log dir. That's not typical of a production
cluster, so this is primarily a test environment change.

The reason for making this change is so that we have the option of
placing files in the log directory before initializing the FsManager.
If we put any files into the log dir before initializing the FsManager
for the first time, which happens in nearly every integration test, the
FsManager will refuse to start up while returning the following error:

  Already present: FSManager root is not empty: /path/to/data/root

In a breakpad-related change following this one, we default the minidump
path to a subdirectory of the log directory. Since breakpad installs
signal handlers designed to collect a form of coredump, and we naturally
initialize breakpad very early in the daemon startup process in order to
collect information on any possible crash, we create the minidump
destination directories before initializing the FsManager. Without
separating the log directory from the data directory, creation of these
subdirectories in the data directory before initializing the FsManager
results in the error described above.

The simplest solution to this problem is to separate the data and log
directories in our integration testing environment.

Change-Id: Iba23c429989df524da51eb012a491766df06e955
Reviewed-on: http://gerrit.cloudera.org:8080/5619
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>